CBOE-6101 (Partial) (#297)
* Fix CBOE-6084: Duplicate Resolution: 'Add Batch' button should be renamed as 'Move Batches' in the duplicate resolution page when updating unique structure to duplicate

* Fix CBOE-6085: Submit New Compound: 'Select Batch' option is not required in Submit New Compound page

* CBOE-5204
Custom Tables-Sequences: Mandatory fields are not specified in 'Add Sequence' page

* Use form to edit custom tables if there are many fields

* CBOE-5204
Custom Tables-Sequences: Mandatory fields are not specified in 'Add Sequence' page

* CBOE-5491
 Register Marked option is not available to register multiple records
- POST api/records/bulk-register - return bulk registration record details

* CBOE-6101 (Partial)
Site specified prefixes are not available in Submit New Compound page

This fix allows users to clear the site field of sequence data.  Without it, once a site value is assigned to sequence, there is no way to clear it.