CBOE-6748 (#775)
Description:
Reg-Inv: Clicking on Request Material opens "Request Sample from an Inventory Container " thrice

Remedy:
Issue is caused by adding click event for handling toggle in the div encapsultaing the Accordion title and the command buttons. Remedy is to remove the click event from the div and add it only to the Accordion title tag.