CBOE-7032 (#952)
Regression: System allows to mark more number of records than the value set at "MarkedHitsMax" setting

Remedy: Do not allow users to mark all when it is already showing marked records.  This is because there is no point of marking already marked records.  When user is in a hit-list of a search, user can select all records within the hit-list as long as the record count is less than the system limit.