CBOE-7140: Assembly version in the application  should be update to 18.1 everywhere (#993)
* CBOE-7324: Grid Management > New support chrome browser
CBOE-7325: Grid Management > Edit and delete support chrome browser.
Fix: Had to include ID for elements with class names, Also had to get rid of display attribute instead used visibility.
Display order had to be modified for look and feel as we used visibility attribute.

* CBOE-7140: Assembly version in the application  should be update to 18.1 everywhere
Fix: As there was an error while running the Inventory DB scripts which was caused as the depth of patches reached 20. Had to stop the patch execution at 17.1.0 and start a new one.

* CBOE-7140: Assembly version in the application  should be update to 18.1 everywhere
Fix: The Changes done here are to make sure the patch after depth 20 are executed.