CBOE-7380-Chrome: Manage Requests- All filter options are not visible in default pop up window (#1021)
* CBOE-7394-Chrome: Unwanted blank white space in container list panel
Remedy - Add class instaed of id for browser compatability and change the top to 60

* CBOE-7394-Chrome: Unwanted blank white space in container list panel
Remedy - Add class instaed of id for browser compatability and change the top to 60

* CBOE-7367-Chrome: Clicked links should be grayed out in Manage Sample Request page
Remedy - Add grey style for the clicked links through css

* CBOE-7380-Chrome: Manage Requests- All filter options are not visible in default pop up window
Remedy - Set a seperate type for Task link from the browse page. So that the menu page will come with correct width.