CBOE-7806: Chrome:Selected Container Admin value is not updating , ap… (#1131)
* CBOE-7806: Chrome:Selected Container Admin value is not updating , application taking time to display the selected value when return back to 'required' tab of container creation page from another tabs
Issue: Before setting the values to the controls there is a browser check option where if the IE version is 6 or more the settimeout will be 1 millisecond otherwise it will be 3 seconds. This is what is causing the actual delay in chrome.
Fix: After making the hold time to 1 millisecond and removing the other browser check it is working as expected.