CBOE-7782: Chrome: Container creation from ACX substance leads to internal Server error (#1155)
Issue: When InLineCdx value is having "data:chemical/x-cdx;base64,0" or "data:chemical/x-cdx;base64," the Instr function value is 0. Because of this Mid function throws an error.
Fix: Had to correct the variable name in CreateSubstance2.asp where we call DispaySubstance function. The variable typo used was wrong. Instead of dBStructure it was dbStructure