CBOE-7123 Chrome Browser:COE-ELN Integration- Additional blank pop up window is displaying along with ELN instance if user try to launch E-Notebook (#1336)
Solution: Set the newWindow attribute value for Utilities link group to false so that a new tab window is opened instead of a new separate window, while opening the target application of the link. The tab window will be closed immediately when the target application is launched or downloaded. However the tab remains to show any errors if any, as it was with new window case. This approach of opening a tab window is chosen since Chrome won't close the new window automatically as it is with IE, after opening/downloading a non browser file.