CBOE-7331-Create Location: Location owner drop down is listing removed users (#1617)
* CBOE-8441-Selected pick list value is not saved in the configuration
Remedy - Added the code for saving the pick list configuration and display the values in the form

* CBOE-8441-Selected pick list value is not saved in the configuration

* CBOE-8441-Selected pick list value is not saved in the configuration
Remedy - Added the code for saving the pick list configuration and display the values in the form

* CBOE-8441-Selected pick list value is not saved in the configuration
Remedy - Added the code for saving the pick list configuration and display the values in the form - Cover all the regression area also

* CBOE-8442-"Choose a default value" drop down is not visible when the user try to modify the value.
Remedy - Added the code for setting the dropdown values for edit mode also

* CBOE-8480-12.1.3-18.1.1: Approve icon is not available in list and details view of temp records in Classic Registration
Remedy - Update the status ID column null values in the temporary_batch table

* CBOE-8520-Rgression: Classic Reg- (Inventory Integration Enable + FullContainerForm=False), unable to create container from registered record.
Remedy - Check the browser and call the appropriate method for each browser

* CBOE-8531-Regression:PickList Sorting order is not working
Remedy - Implement the code for including the sort order in the sql query

* CBOE-8531-Regression:PickList Sorting order is not working
Remedy - Removed the unwanted code from the file.

* CBOE-8521-WebReg- Getting Destination Require validation message even destination selected.
Remedy - Form control LocationID not return the proper location values in chrome browser, so used session variable for store and retrieve the correct location ID

* CBOE-8524-Classic Reg:-Registration/InventoryEnabled- Few fields are hided with default pop up window and Unwanted scroll bars in create container page
Remedy - Increase the width of the pop up window for display all controls.

* CBOE-7458-New link is not become grey by default or after click on it in Manage Requests page
Remedy - Requesttype is null so it set as Pending that should have disable1 for setting the New link as grayed

* CBOE-7788-Type the location barcode and press 'Tab' not updating location name in Delete Location page
Remedy - Added missing input tags with the form

* Revert "CBOE-7788-Type the location barcode and press 'Tab' not updating location name in Delete Location page"

This reverts commit 5d0ae1cd79789115248b623fe7277c0d0a6475ed.

* CBOE-7788-Type the location barcode and press 'Tab' not updating location name in Delete Location page
Remedy - Added missing input tags with the form

* CBOE-8613-Chrome: Unable to add deleted user again and all fields are showing blank including available Roles
Remedy -  "document.attachEvent" is not supported in Chrome browser. So i changed it on condition basis and also update the javascript for display the validation message properly.

* CBOE-8522-Registration/InventoryEnabled:- Application should not allow to create container at Root(0) location
Remedy - added the validation in the javascript

* CBOE-8762-After inactivating a prefix AB from cboe page the prefix AB is still showing in ELN Registration Preference Prefix dropdown
Remedy - Default prefix updation will work only with the Add mode of the Formgroup display mode. So I changed it.

* CBOE-8765-ChemACX Database radio button options needs to be removed from Server configuration tool

* CBOE-7331-Create Location: Location owner drop down is listing removed users
Remedy - Implement the filter for removing inactive users from the dropdown